ggswissmaps: Offers Various Swiss Maps as Data Frames and 'ggplot2' Objects

Offers various swiss maps as data frames and 'ggplot2' objects and gives the possibility to add layers of data on the maps. Data are publicly available from the swiss federal statistical office. In addition to the maps2 object (a list of 8 swiss maps, at various levels), there are the data frames with the boundaries used to produce these maps (shp_df, a list with 8 data frames).

See also

Author

Maintainer: Sandro Burri gibo.gaf@gmail.com

Authors:

Eric Stephani eric.stephani@bluewin.ch

Other contributors:

Lisa Bottinelli lisa.bottinelli@gmail.com [contributor]