A list with 8 data frames of swiss territory boundaries, at various levels.

Every element of the list is a data frame, which can be used with 'ggplot2'.

Format

A list with 8 data frames with swiss territory boundaries (at various levels).

Source

https://data.geo.admin.ch/browser/index.html#/collections/ch.bfs.historisierte-administrative_grenzen_g1?.language=en. Download date: 2015-08-18

Details

Columns are not all the same across data frames, but usually they have the following in common:

long. Longitude coordinate (x)

lat. Latitude coordinate (y)

group. A factor to be used to plot the polygons correctly (with ggplot2)

Examples

data(shp_df)
class(shp_df)
#> [1] "list"
length(shp_df)
#> [1] 8
names(shp_df)
#> [1] "g1b15"      "g1g15_encl" "g1g15_li"   "g1g15"      "g1k15"     
#> [6] "g1l15"      "g1r15"      "g1s15"     
str(shp_df[["g1k15"]])
#> 'data.frame':	9216 obs. of  20 variables:
#>  $ long   : int  692429 692993 693163 693048 693243 693606 693763 693872 694800 695433 ...
#>  $ lat    : int  281173 280860 280421 280201 279410 278826 278851 278450 278137 277578 ...
#>  $ order  : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ hole   : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ piece  : Factor w/ 6 levels "1","2","3","4",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ group  : Factor w/ 56 levels "0.1","1.1","1.2",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ id     : chr  "0" "0" "0" "0" ...
#>  $ KTNR   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ GRNR   : int  4 4 4 4 4 4 4 4 4 4 ...
#>  $ AREA_HA: int  172895 172895 172895 172895 172895 172895 172895 172895 172895 172895 ...
#>  $ X_MIN  : int  669244 669244 669244 669244 669244 669244 669244 669244 669244 669244 ...
#>  $ X_MAX  : int  716900 716900 716900 716900 716900 716900 716900 716900 716900 716900 ...
#>  $ Y_MIN  : int  223896 223896 223896 223896 223896 223896 223896 223896 223896 223896 ...
#>  $ Y_MAX  : int  283343 283343 283343 283343 283343 283343 283343 283343 283343 283343 ...
#>  $ X_CNTR : int  691800 691800 691800 691800 691800 691800 691800 691800 691800 691800 ...
#>  $ Y_CNTR : int  252000 252000 252000 252000 252000 252000 252000 252000 252000 252000 ...
#>  $ Z_MIN  : int  331 331 331 331 331 331 331 331 331 331 ...
#>  $ Z_MAX  : int  1290 1290 1290 1290 1290 1290 1290 1290 1290 1290 ...
#>  $ Z_AVG  : int  533 533 533 533 533 533 533 533 533 533 ...
#>  $ Z_MED  : int  504 504 504 504 504 504 504 504 504 504 ...